grpc-py/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
grpcio==1.62.1
    # via grpcio-tools
grpcio-tools==1.62.1
mysql-connector-python==8.3.0
protobuf==4.25.3
    # via grpcio-tools
setuptools==69.2.0
    # via grpcio-tools
